requirements.txt
pandas
mistralai
python-dotenv